Move code related to indoo.rs to LocationService Refactor mainActivity, by moving all integration code to a new class LocationService. This service implements IndoorsLocationListner. All events related to location will therefore be relayed to the LocationService
