+++ app/src/main/java/no/ntnu/tpg4850/frugalar/MainActivity.java
@@ -42,7 +42,7 @@ import java.util.List;
 /**
  * Cardboard application. Will display a camera preview for each eye.
  */
-public class MainActivity extends CardboardActivity implements CardboardView.StereoRenderer, OnFrameAvailableListener, IndoorsLocationListener {
+public class MainActivity extends CardboardActivity implements CardboardView.StereoRenderer, OnFrameAvailableListener {
 
     private static final String TAG = "MainActivity";
 
@@ -52,7 +52,7 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     private CardboardOverlayView mOverlayView;
     private CardboardView cardboardView;
     private SurfaceTexture surface;
-    private IndoorsSurfaceFragment indoorsFragment;
+    private LocationService location;
     //private float[] mView;
     //private float[] mCamera;
     private CameraEyeTransformer cameraPreviewTransformer;
@@ -87,17 +87,8 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        IndoorsFactory.Builder indoorsBuilder = new IndoorsFactory.Builder();
-        IndoorsSurfaceFactory.Builder surfaceBuilder = new IndoorsSurfaceFactory.Builder();
-        indoorsBuilder.setContext(this);
-        indoorsBuilder.setApiKey("8fe0cc81-2098-41cf-b8c5-9fa5874ef8a6");
-        indoorsBuilder.setBuildingId((long) 272563595);
-        indoorsBuilder.setUserInteractionListener(this);
-        surfaceBuilder.setIndoorsBuilder(indoorsBuilder);
+        location = new LocationService(this);
 
-        indoorsFragment = surfaceBuilder.build();
-
-        //FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
 
         this.cameraPreviewTransformer = new CameraEyeTransformer();
         setContentView(R.layout.common_ui);
@@ -188,46 +179,5 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     public void onCardboardTrigger() {
         //When magnetic button has been triggered
         Log.i(TAG, "onCardboardTrigger");
-    }
-
-    @Override
-    public void loadingBuilding(int i) {
-        Log.i(TAG, "Loading Building");
-
-    }
-
-    @Override
-    public void buildingLoaded(Building building) {
-        Log.i(TAG, "Loaded Building");
-    }
-
-    @Override
-    public void leftBuilding(Building building) {
-        Log.i(TAG, "Left building");
-    }
-
-    @Override
-    public void positionUpdated(Coordinate coordinate, int i) {
-        Log.i(TAG, "position update" + coordinate);
-    }
-
-    @Override
-    public void orientationUpdated(float v) {
-        Log.i(TAG, "Orientation");
-    }
-
-    @Override
-    public void changedFloor(int i, String s) {
-        Log.i(TAG, "Changed floor");
-    }
-
-    @Override
-    public void enteredZones(List<Zone> zones) {
-        Log.i(TAG, "Entered zone");
-    }
-
-    @Override
-    public void onError(IndoorsException e) {
-        Log.i(TAG, "Error");
     }
 }
